Add support for Guid type
The Guid type in .NET logically maps the the AVRO UUID type.

See https://avro.apache.org/docs/1.11.1/specification/#uuid for
reference and
https://github.com/apache/avro/blob/7c17fd534b477236a9a1b635a5375c99ce811e94/lang/csharp/src/apache/test/AvroGen/AvroGenSchemaTests.cs#L691
for a c# reference from upstream avro generation.

As this is a simple usage of string with a logical type it is pretty
straight forward to add.